Initial commit
Start of with the same project structure as the Pyramid starter project, with the following changes:

Removed all PasteDeploy style startup, it feels really old. I have used it before but in 2024 it just feels a bit dated.

Instead the startup is in __main__.py and `app` is in __init__.py, you can start a development server using Waitress by just running:

    python -m sambal

And if you want to start the app in production mode with gunicorn, just pip install gunicorn and run it like this:

    gunicorn sambal:app -w 4

That feels a lot tidier and I prefer this much more to starting with pserve and ini files.

In addition, all old style packaging (setup.py, MANIFEST.in) has been removed and replaced with a pyproject.toml using a modern project layout.

It doesn't use Poetry because there are very little dependencies, a lockfile isn't necessary at this stage.