views: add sample query to home view for fetching User records
It doesn't need to display these users in the jinja2 template, it's just there so that something is using the request.samdb property for testing purposes.

Note that users is a Query object which is iterable, a class that can be used like a generator for returning records when looping over it.

The fact that the import is still under "samba.netcmd.domain.models" is something I want to change to "samba.domain.models" in upstream Samba, as the models have nothing to do with the cli commands.