auth: add login and logout pages
For now the SambalSecurityPolicy.permits method just checks if request.identity is not None which is essentially "is the user logged in" regardless of the permission name used on the home view.

I wanted to push the login form first while it was in a working state, then work out ACLHelper separately.

Uses WTForms, the form is unstyled and no attention has been put towards how it looks at this point, that comes later.