tweens: add security header tween
Uses a single tween because bringing in a library like pyramid-secure-response when we need strictly local and same-site based settings anyway is overkill.

Takes care of all the headers except HSTS headers, doing it in a single tween for simplicity.

The HSTS headers can be done after as they need to be put behind a flag (env var) because in development you can't have HSTS turned on.