login: bubble up LdbError so that it can be handled better
Step one is to change connect_samdb, also placing host as the first argument for consistency.

connect_samdb now expects host, username and password to be strings and won't do checking itself anymore if they are present.

It will no longer catch LdbError in connect_samdb either and just let it raise instead. This is so that it can be handled further up the chain.

In get_samdb, which provides the request.samdb property, we keep the return type as Optional[SamDB], but only return None if there are no credentials in the session at all (it raises KeyError).

If the credentials in the session worked before, but now no longer work, for example server is gone, just let it raise LdbError in get_samdb because we need to know at this point when that happens rather than return None (which means not authenticated).