models: add support for adding newer models in the sambal source
Make it possible to have models in Sambal before they have been added to upstream Samba.

Models can always be donated upstream to Samba over time. As the licenses are identical there is no problem.

This just makes it quicker to get lots of models committed and tested before they make it upstream (or not).

Also, this make it possible to override models already in Samba or add new fields that aren't yet mapped in Samba (see OrganizationalUnit as an example).